mirador

Module assets

Assets Module

This module contains all game assets embedded in the binary using include_bytes!(). This ensures that all assets are available at runtime without requiring external files.

Constants

AUDIO_BEEPER_RISE

Beeper rise sound effect data

AUDIO_COMPLETE

Level complete sound effect data

AUDIO_JEFFREE_STAR_ASMR

Jeffree Star ASMR sound effect data

AUDIO_SELECT

Menu select sound effect data

AUDIO_SINGLE_STEP

Single step sound effect data

AUDIO_SLIME_TRACK

Slime track sound effect data

AUDIO_UPGRADE

Upgrade sound effect data

AUDIO_WALL

Wall hit sound effect data

BLANK_ICON

Blank icon image data

COMPASS_BASE

Compass base image data

DASH_ICON

Dash ability icon image data

FRANKIE_IMAGE

Frankie character image data

GOLD_COMPASS

Gold compass image data

HANKEN_GROTESK_BOLD

Hanken Grotesk Bold font data

HANKEN_GROTESK_MEDIUM

Hanken Grotesk Medium font data

HANKEN_GROTESK_REGULAR

Hanken Grotesk Regular font data

HEAD_START_ICON

Head start ability icon image data

JEFFREE_IMAGE

Jeffree character image data

MAZE_ICON_IMAGE

Maze icon image data

MUSIC_MAIN_TRACK

Main game music track data

MUSIC_MAIN_TRACK_STRIPPED

Stripped main game music track data

NEEDLE_0

Compass needle position 0 image data

NEEDLE_1

Compass needle position 1 image data

NEEDLE_2

Compass needle position 2 image data

NEEDLE_3

Compass needle position 3 image data

NEEDLE_4

Compass needle position 4 image data

NEEDLE_5

Compass needle position 5 image data

NEEDLE_6

Compass needle position 6 image data

NEEDLE_7

Compass needle position 7 image data

NEEDLE_8

Compass needle position 8 image data

NEEDLE_9

Compass needle position 9 image data

NEEDLE_10

Compass needle position 10 image data

NEEDLE_11

Compass needle position 11 image data

SILENT_STEP_ICON

Silent step ability icon image data

SLIME_IMAGE

Slime enemy image data

SLOWER_SECONDS_ICON

Slower seconds ability icon image data

SPEED_UP_ICON

Speed up ability icon image data

TALL_BOOTS_ICON

Tall boots ability icon image data

TILES_IMAGE

Tiles texture image data

TITLE_IMAGE

Game title image data

UNKNOWN_ICON

Unknown ability icon image data

Functions

compass_needles

Returns all compass needle textures in order

fonts

Returns all font data with their names

icon_textures

Returns all icon textures with their IDs