mirador::math

Module coordinates

Coordinate system transformations for the maze.

This module provides utilities to convert between different coordinate systems used in the game:

Maze Grid Coordinates: Used for maze generation (rows/columns)

World Coordinates: 3D space where the player moves (x, y, z)

It centralizes all coordinate transformations and provides utilities for finding special cells like the entrance (bottom left) and exit.

Modules

constants

Constants for special positions in the maze

positions 🔒

Special positions and cell finding utilities.

transformations 🔒

Coordinate system transformations between different representations.

Enums

Direction

Enum representing cardinal directions in the maze

Functions

calculate_cell_size

Calculates the size of a single cell in world units.

direction_to_yaw

Translates a direction into a yaw angle.

get_adjacent_cell

Gets the cell adjacent to the given cell in the specified direction.

get_bottom_left_cell

Gets the bottom-left cell of the maze.

get_bottom_left_cell_position

Calculates the world coordinates of the bottom-left cell in the maze.

get_bottom_right_cell

Gets the bottom-right cell of the maze.

get_direction_between_cells

Gets the relative direction from one cell to another.

get_top_left_cell

Gets the top-left cell of the maze.

get_top_right_cell

Gets the top-right cell of the maze.

maze_to_world

Converts a maze grid cell to world coordinates.

wall_grid_to_maze_cell

Converts a position in the maze wall grid to a position in the maze cell grid.

world_to_maze

Converts 3D world coordinates to a maze grid cell.

yaw_to_direction

Converts a yaw angle (in degrees) to a cardinal direction.