mirador::maze::generator

Struct MazeGenerator

pub struct MazeGenerator {
    pub maze: Arc<Mutex<Maze>>,
    union_find: UnionFind,
    edges: Vec<Edge>,
    current_edge: usize,
    pub generation_complete: bool,
    pub connected_cells: HashSet<Cell>,
    pub fast_threshold: usize,
    pub fast_mode: bool,
}

Maze generator using Kruskal’s algorithm

Fields
maze: Arc<Mutex<Maze>>

The maze being generated (wrapped in Arc for thread safety)

union_find: UnionFindedges: Vec<Edge>current_edge: usizegeneration_complete: bool

Indicates if generation is complete

connected_cells: HashSet<Cell>

Set of cells currently connected in the maze

fast_threshold: usize

Number of edges remaining when we switch to fast mode

fast_mode: bool

Whether we’re in fast mode

Implementations

impl MazeGenerator

pub fn new(width: usize, height: usize) -> (Self, Arc<Mutex<Maze>>)

Creates a new maze generator with the given dimensions Returns both the generator and a shared reference to the maze

pub fn step(&mut self) -> bool

Performs one step of maze generation Returns true if a wall was removed in this step

pub fn is_complete(&self) -> bool

Checks if maze generation is complete

pub fn get_progress(&self) -> (usize, usize)

Returns the current progress of generation (processed edges, total edges)

pub fn get_progress_ratio(&self) -> f32

Returns the generation progress as a ratio (0.0 to 1.0)